+++ lib/test/clients/sequencer/SequencerWsService.unit.test.ts
@@ -1,12 +1,15 @@
 import { mocked } from 'ts-jest/utils'
 import { GatewaySequencerCommand } from '../../../src/clients/gateway/models/Gateway'
 import type * as Req from '../../../src/clients/sequencer/models/PostCommand'
+import type { SequencerStateResponse } from '../../../src/clients/sequencer/models/SequencerRes'
 import {
   QueryFinal,
-  SequencerWebsocketRequest
+  SequencerWebsocketRequest,
+  SubscribeSequencerState
 } from '../../../src/clients/sequencer/models/WsCommand'
 import { SequencerServiceImpl } from '../../../src/clients/sequencer/SequencerServiceImpl'
 import { SubmitResponseD } from '../../../src/decoders/CommandDecoders'
+import { SequencerStateResponseD } from '../../../src/decoders/SequencerDecoders'
 import { ComponentId, Prefix } from '../../../src/models'
 import { HttpTransport } from '../../../src/utils/HttpTransport'
 import { Ws } from '../../../src/utils/Ws'
@@ -30,5 +33,15 @@ test('SequencerService should receive submit response on query final using webso
   verify(mockWs.singleResponse).toBeCalledWith(
     new GatewaySequencerCommand(componentId, new QueryFinal('12345', 1000)),
     SubmitResponseD
+  )
+})
+test('SequencerService should receive sequencer state response on subscribe sequencer state using websocket | ESW-307', async () => {
+  const callback = () => ({})
+  await sequencer.subscribeSequencerState()(callback)
+
+  verify(mockWs.subscribe).toBeCalledWith(
+    new GatewaySequencerCommand(componentId, new SubscribeSequencerState()),
+    callback,
+    SequencerStateResponseD
   )
 })